Call pthread's recursive mutex on Linux (issue #245), part 1
Replace "#ifdef PTHREAD_MUTEX_RECURSIVE"
   with "#ifdef HAVE_PTHREAD_MUTEX_RECURSIVE"
and #define HAVE_PTHREAD_MUTEX_RECURSIVE in config.h (unconditionally)
so the recursive mutex code is used if PTHREAD_MUTEX_RECURSIVE is an
enum and not "defined" (as a macro).

Note: the added compiler warnings "Compiling with (or w/o) ..."
  must be removed after testing.

To do: part 2 should add CMake and configure checks.